# core/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Command-and-control core backend API.",
        "title": "Logos Core API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/auth/oidc/callback": {
            "get": {
                "description": "Provider redirect target. On success sets the usual auth cookies and redirects into the SPA; on failure redirects to the SPA login page with ?error=\u003ccode\u003e.",
                "tags": [
                    "Auth"
                ],
                "summary": "Single sign-on callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization code",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "State echoed by the provider",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Provider-side error",
                        "name": "error",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    }
                }
            }
        },
        "/auth/oidc/login": {
            "get": {
                "description": "Seals a handshake (state, nonce, PKCE verifier) into a short-lived httpOnly cookie and redirects the browser to the OpenID provider. Optional return_to is a SPA-relative path to land on afterwards.",
                "tags": [
                    "Auth"
                ],
                "summary": "Start single sign-on",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SPA-relative path to return to after login",
                        "name": "return_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/login": {
            "post": {
                "description": "Authenticate with username and password. Tokens are set as httpOnly cookies; CSRF token in a non-httpOnly cookie.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SessionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requests.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "responses.SessionResponse": {
            "type": "object",
            "properties": {
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer JWT token (e.g. \"Bearer eyJhb...\")",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}